+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [Fixed PDF export on Windows](https://github.com/DigiScore/neoscore/issues/37)
 - Updated `neoscore.render_image()` to block by default. Set the new kwarg `wait=False` to finalize export asynchronously.
+- Fixed bug where image and PDF export exceptions could be ignored. Exceptions in image export threads now propagate on join via the new `PropagatingThread` class.
 
 # 0.1.0 (2020-05-15)